fix warning

+++ include/sprite2/Anim2Curr.h
@@ -29,7 +29,7 @@ public:
 	void SetStaticCurrFrame(int frame);
 
 	uint64_t* GetFramePtr(int idx) {
-		if (idx >= 0 && idx < m_frames_ptr.size()) {
+		if (idx >= 0 && idx < static_cast<int>(m_frames_ptr.size())) {
 			return &m_frames_ptr[idx];
 		} else {
 			return NULL;
